+++ src/main/java/com/example/Lion.java
@@ -32,22 +32,3 @@ public class Lion {
         return feline.getFood("Хищник");
     }
 }
-//    public Lion(String sex) throws Exception {
-//        if ("Самец".equals(sex)) {
-//            hasMane = true;
-//        } else if ("Самка".equals(sex)) {
-//            hasMane = false;
-//        } else {
-//            throw new Exception("Используйте допустимые значения пола животного - самец или самка");
-//        }
-//    }
-//    Feline feline = new Feline();
-//    public int getKittens() {
-//        return feline.getKittens();
-//    }
-//    public boolean doesHaveMane() {
-//        return hasMane;
-//    }
-//    public List<String> getFood() throws Exception {
-//        return feline.getFood("Хищник");
-//    }
